Rename AssertionTests to GeneralAssertionTests in contrast to nearby specific assertion test classes.
